pyproject.toml
[project]
name = "paper-ocr"
version = "0.1.0"
description = "Born-digital-aware PDF to Markdown pipeline using olmOCR via DeepInfra"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "paper_ocr"}]
dependencies = [
  "pymupdf>=1.24.0",
  "Pillow>=10.0.0",
  "pyyaml>=6.0.1",
  "python-dotenv>=1.0.1",
  "openai>=1.40.0",
  "pytest>=8.0.0",
]

[project.scripts]
paper-ocr = "paper_ocr.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
